Support git-like command names
This commit adds support for git-like parsing of command names in an
effort to improve Cog's CLI usability. Concretely, command names
containing hyphens are inferred from user input as long as the head of
the arg list is a string.

Let's work thru an example to see how this works. Assume we have a EC2
bundle installed which contains a command named "instances-list". Prior
to this change the only ways to invoke the command would be for the user
to refer to it as "ec2:instances-list" or "instances-list".

This change continues to support the above references and adds a third,
subcommand-like reference: "ec2:instances list". This works for any
number of hyphens so long as the head of the arg list is a string. Taken
to an extreme, the command "ec2:instances-list-and-purge" can be
referenced in all the following ways: "ec2:instances-list-and-purge",
"instances-list-and-purge", "ec2:instances list and purge", "instances
list and purge", etc.

Parsing hyphenated commands as subcommands is achieved by modifying the
logic used by `Piper.Command.Parser` to resolve commands during parse
time. The revised logic looks like this:

0. Call the resolver with the current bundle and command name.
1. If the command name is found, stop and return the pipeline (in case
of expanded aliases) or bundle and commmand (in case of commands).
2. If the first arg is a string:
  2a.  Create a new command name by concatenating the first arg onto
  the command name separated by a hyphen.
  2b. Goto 0.
3. If the first arg is not a string, return :not_found.